Proposed Update to Branch Usage
This is a proposal to change the way we use branches, to ensure that master is always in a stable state, without needing to worry about breaking Docker. This will also allow us to release PyPi packages that provide the same features as Docker, without flooding the system with each commit.